+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Move pagination to datagouv-components [#365](https://github.com/etalab/udata-front/pull/365)
 - Fix release CI [#380](https://github.com/etalab/udata-front/pull/380)
+- Add stories to resource component [#364](https://github.com/etalab/udata-front/pull/364)
 
 ## 3.5.4 (2024-03-20)